requirements/publisher.txt
# SHA1:e2360a7388b9f6b7b29adbd471c7656d10aa5064
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
paho-mqtt==1.6.1
    # via -r requirements/publisher.in
